chore:stopbug

+++ latest-global/src/main/java/com/github/jartisan/latest/global/utils/UnitUtil.java
@@ -17,7 +17,7 @@ public class UnitUtil {
         int flooring, numlength;
         String[] unitname = {"B", "kB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"};
         numlength = String.valueOf(num).length();
-        flooring = (int) Math.floor((numlength - 1) / 3);
+        flooring = (int) Math.floor((numlength - 1) / (double)3);
         String size = unitname[flooring];
         java.text.DecimalFormat dfmt = new java.text.DecimalFormat("###.## " + size);
         return dfmt.format((double) num / Math.pow(1024, flooring));
